refactor move teamTrials directory to src/components
