+++ src/components/teamTrials/TeamTrialBanner.test.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from "react";
-import TeamTrialBanner from "@/extensionConsole/pages/teamTrials/TeamTrialBanner";
+import TeamTrialBanner from "@/components/teamTrials/TeamTrialBanner";
 import { render, screen, waitFor } from "@/extensionConsole/testHelpers";
 import { organizationResponseFactory } from "@/testUtils/factories/organizationFactories";
 import { type Timestamp } from "@/types/stringTypes";
