+++ src/components/teamTrials/TrialAwareButton.tsx
@@ -15,11 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import TrialCallToActionLink from "@/extensionConsole/pages/teamTrials/TrialCallToActionLink";
-import useGetAllTeamScopes from "@/extensionConsole/pages/teamTrials/useGetAllTeamScopes";
+import TrialCallToActionLink from "@/components/teamTrials/TrialCallToActionLink";
+import useGetAllTeamScopes from "@/components/teamTrials/useGetAllTeamScopes";
 import useTeamTrialStatus, {
   TeamTrialStatus,
-} from "@/extensionConsole/pages/teamTrials/useTeamTrialStatus";
+} from "@/components/teamTrials/useTeamTrialStatus";
 import { type RegistryId } from "@/types/registryTypes";
 import { type IconProp } from "@fortawesome/fontawesome-svg-core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
