+++ src/extensionConsole/App.tsx
@@ -56,7 +56,7 @@ import DatabaseUnresponsiveBanner from "@/components/DatabaseUnresponsiveBanner"
 import ActivateModPage from "@/extensionConsole/pages/activateMod/ActivateModPage";
 import { RestrictedFeatures } from "@/auth/featureFlags";
 import { useLocation } from "react-router";
-import TeamTrialBanner from "@/extensionConsole/pages/teamTrials/TeamTrialBanner";
+import TeamTrialBanner from "@/components/teamTrials/TeamTrialBanner";
 
 // Register the built-in bricks
 registerEditors();
