+++ src/sidebar/SidebarBody.tsx
@@ -22,6 +22,7 @@ import RestrictedUrlPopupApp from "@/tinyPages/RestrictedUrlPopupApp";
 import useConnectedTargetUrl from "./hooks/useConnectedTargetUrl";
 import { getReasonByUrl as getRestrictedReasonByUrl } from "@/tinyPages/restrictedUrlPopupUtils";
 import DatabaseUnresponsiveBanner from "@/components/DatabaseUnresponsiveBanner";
+import TeamTrialBanner from "@/components/teamTrials/TeamTrialBanner";
 
 const SidebarReady: React.FC<{ url: string }> = ({ url }) => {
   const restricted = getRestrictedReasonByUrl(url);
@@ -40,6 +41,7 @@ function SidebarBody() {
   return (
     <>
       <Header />
+      <TeamTrialBanner />
       <DatabaseUnresponsiveBanner />
       <div className="full-height">{url && <SidebarReady url={url} />}</div>
     </>
